requirements.txt
PyYAML==6.0.1
requests==2.31.0
BeautifulSoup4==4.12.3
geopy==2.4.1 